Move source to lib/ folder.
