+++ js/package.json
@@ -18,7 +18,7 @@
   },
   "author": "Microsoft Corporation",
   "license": "Apache-2.0",
-  "main": "edge.js",
+  "main": "./lib/edge.js",
   "directories": {
     "test": "test"
   },
